Show pair count instead of row count
